Ethereum TVL refactoring
